Conflicts:
	Projects/TestBed/Classes/Infrastructure/GameCore.cpp
	Sources/Internal/UI/UITextFieldiPhone.mm

Former-commit-id: 632d1d5e2f717c53ff83dd811ef34d785717f5db

+++ Sources/Internal/Render/2D/TextLayout.cpp
@@ -94,8 +94,8 @@ void TextLayout::PrepareCharSizes()
     DVASSERT(characterSizes.size() == preparedText.length());
     if (useBiDi)
     {
-        uint32 size = characterSizes.size();
-        for (uint32 i = 0; i < size; ++i)
+        size_t size = characterSizes.size();
+        for (size_t i = 0; i < size; ++i)
         {
             if (bidiHelper.IsBiDiSpecialCharacter(preparedText[i]))
             {
